BACKEND/requirements.txt
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1671137333039/work
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
crispy-tailwind==0.5.0
cryptography==39.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.6
django-allauth==0.52.0
django-appconf==1.0.5
django-countries==7.5.1
django-filter==22.1
django-guardian==2.4.0
django-htmx==1.13.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
Markdown==3.4.1
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1669306857274/work
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
yahoofinancials==1.13
zipp==3.13.0
zope.interface==5.5.2
